requirements.txt
# API
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.11.5
pydantic-settings==2.9.1
python-dotenv==1.1.0
email-validator>=2.0.0
python-jose


# MongoDB
motor==3.7.1
pymongo==4.13.0

# Desarrollo
pre-commit==3.6.2
